[LTO][ELF] Add --lto-exit-on option
Allows the linker to "exit early" after running specific LTO module hooks.

This is useful in combination with --save-temps= as it gives the flexibility of stopping lld once the desired files are emitted. This minimizes wasted time (from performing optimizations, codegen, linking, etc) when all that is desired from the linker is the "--save-temps=import" output (which is completed before the aforementioned tasks), as in the case of extracting a training corpus for MLGO.

Differential Revision: https://reviews.llvm.org/D127779